requirements.txt
python-craigslist
boto3